#20 로그아웃 및 redux 파일 세분화
