+++ src/main.jsx
@@ -6,18 +6,18 @@ import { lightTheme } from "@/styles/theme";
 import GlobalStyles from "@/styles/GlobalStyles";
 import ReactModal from "react-modal";
 import ReactDOM from "react-dom/client";
-import { store } from "@/store/store";
+import { store } from "@/store";
 
 if (ReactModal && ReactModal.setAppElement) {
-  ReactModal.setAppElement("#root");
+	ReactModal.setAppElement("#root");
 }
 ReactDOM.createRoot(document.getElementById("root")).render(
-  <Provider store={store}>
-    <GoogleOAuthProvider clientId="379597382111-vo2ht0r8a3d0ais7v12q7777lu48al1a.apps.googleusercontent.com">
-      <ThemeProvider theme={lightTheme}>
-        <GlobalStyles />
-        <App />
-      </ThemeProvider>
-    </GoogleOAuthProvider>
-  </Provider>
+	<Provider store={store}>
+		<GoogleOAuthProvider clientId="379597382111-vo2ht0r8a3d0ais7v12q7777lu48al1a.apps.googleusercontent.com">
+			<ThemeProvider theme={lightTheme}>
+				<GlobalStyles />
+				<App />
+			</ThemeProvider>
+		</GoogleOAuthProvider>
+	</Provider>,
 );
