+++ src/store/index.js
@@ -1,8 +1,8 @@
 import { configureStore } from "@reduxjs/toolkit";
-import userSlice from "./user/user-slice";
+import userSlice from "@/features/user/user-slice.js";
 
 export const store = configureStore({
-  reducer: {
-    user: userSlice,
-  },
+	reducer: {
+		user: userSlice,
+	},
 });
